requirements.txt
pandas==2.2.2
numpy==1.26.4
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
kaggle==1.6.14
